+++ src/Stratis.FederatedPeg.Features.FederationGateway/GlobalSuppressions.cs
@@ -1,0 +1,10 @@
+﻿// This file is used by Code Analysis to maintain SuppressMessage
+// attributes that are applied to this project.
+// Project-level suppressions either have no target or are given
+// a specific target and scoped to a namespace, type, member, etc.
+
+[assembly: System.Diagnostics.CodeAnalysis.SuppressMessage("StyleCop.CSharp.LayoutRules", "SA1503:Braces must not be omitted", Justification = "<Pending>", Scope = "member", Target = "~M:Stratis.FederatedPeg.Features.FederationGateway.MonitorChain.SourceChainSessionManager.RunSessionAsync~System.Threading.Tasks.Task")]
+[assembly: System.Diagnostics.CodeAnalysis.SuppressMessage("Style", "IDE0008:Use explicit type", Justification = "<Pending>", Scope = "member", Target = "~M:Stratis.FederatedPeg.Features.FederationGateway.MonitorChain.Auditors.JsonTransferStore.AddCrossChainTransactionId(NBitcoin.uint256,NBitcoin.uint256)")]
+[assembly: System.Diagnostics.CodeAnalysis.SuppressMessage("StyleCop.CSharp.LayoutRules", "SA1503:Braces must not be omitted", Justification = "<Pending>", Scope = "member", Target = "~M:Stratis.FederatedPeg.Features.FederationGateway.SourceChain.SourceChainSessionManager.RunSessionAsync~System.Threading.Tasks.Task")]
+[assembly: System.Diagnostics.CodeAnalysis.SuppressMessage("StyleCop.CSharp.LayoutRules", "SA1516:Elements must be separated by blank line", Justification = "<Pending>", Scope = "member", Target = "~P:Stratis.FederatedPeg.Features.FederationGateway.TargetChain.CounterChainSession.BlockHeight")]
+
