+++ src/Stratis.FederatedPeg.Features.FederationGateway/Models/ImportMemberKeyRequest.cs
@@ -1,0 +1,14 @@
+﻿using System.ComponentModel.DataAnnotations;
+using Stratis.Bitcoin.Features.Wallet.Models;
+
+namespace Stratis.FederatedPeg.Features.FederationGateway.Models
+{
+    public class ImportMemberKeyRequest : RequestModel
+    {
+        [Required(ErrorMessage = "A mnemonic is required.")]
+        public string Mnemonic { get; set; }
+
+        [Required(ErrorMessage = "A password is required.")]
+        public string Password { get; set; }
+    }
+}
