+++ src/Stratis.FederatedPeg.Features.FederationGateway/SourceChain/ITransferSourceManager.cs
@@ -1,0 +1,13 @@
+﻿using System;
+
+namespace Stratis.FederatedPeg.Features.FederationGateway.SourceChain
+{
+    public interface ITransferSourceManager : IDisposable
+    {
+        void Initialize();
+
+        void Register(TransferLeaderSelector monitorSession);
+
+        void CreateSessionOnCounterChain(int apiPort, TransferLeaderSelector monitorChainSession);
+    }
+}
